SumitMewada_EIR/Sumit_EIR_TaskC_ViewerCrawler/viewer_intelligence/requirements.txt
# YouTube Data Collection
google-api-python-client==2.108.0
yt-dlp==2023.11.16
pytube==15.0.0

# NLP & Language Processing
spacy==3.7.4
langdetect==1.0.9
fasttext-wheel==0.9.2
geotext==0.4.0
pydantic==2.5.3

# Geocoding & Geography
geopy==2.4.1
pycountry==23.12.11

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
folium==0.15.1

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
